Added layout blocks to server views
