+++ server/views/index.hbs
@@ -5,4 +5,6 @@
 
 <div id="root"></div>
 
+{{#forBlock 'scripts'}}
 <script src="{{asset '/build/index.js'}}" type="text/javascript" charset="utf-8"></script>
+{{/forBlock}}
