+++ server/views/layout.hbs
@@ -5,10 +5,12 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>{{title}}</title>
-    <link rel='stylesheet' href='/stylesheets/style.css' />
-    <script src="{{asset '/build/common.js'}}" type="text/javascript" charset="utf-8"></script>
+    <link rel="stylesheet" href="/stylesheets/style.css" />
+    {{{block 'stylesheets'}}}
   </head>
   <body>
     {{{body}}}
+    <script src="{{asset '/build/common.js'}}" type="text/javascript" charset="utf-8"></script>
+    {{{block 'scripts'}}}
   </body>
 </html>
